Always Set Project Dir as Working Dir
Replaces any calls to pwd with a routine that
sets baseDir to the location of any affected
invoked script. This change avoids changes to
directories outside of this project when its
scripts are invoked from outside of the project
directory.

Currently only follows a single symlink since
MacOS is missing the -f flag necessary for
following multiple links.